Clean Up;

+++ NeuralNetworkTest/include/mnist/mnist_reader_common.hpp
@@ -43,7 +43,7 @@ inline std::unique_ptr<char[]> read_mnist_file(const std::string& path, uint32_t
     }
 
     auto size = file.tellg();
-    std::unique_ptr<char[]> buffer(new char[size]);
+	std::unique_ptr<char[]> buffer(new char[(unsigned int)size]);
 
     //Read the entire file at once
     file.seekg(0, std::ios::beg);
